Docs: 파일 구조 통일
