+++ CaffeineDrop/components/eventPage/Card_Brazil.js
@@ -1,12 +1,12 @@
 import React from "react";
 import { Image } from "react-native";
 import styled from "styled-components/native";
-import EthiopiaPng from "../assets/EventPage/Africa1.png";
+import BrazilPng from "../../assets/EventPage/SouthAmerica1.png";
 import {
   responsiveFontSize,
   responsiveWidth,
   responsiveHeight,
-} from "../utils/responsive";
+} from "../../utils/responsive";
 
 import { Dimensions } from "react-native";
 
@@ -16,8 +16,8 @@ const { width, height } = Dimensions.get("window");
 // 태블릿 판별 기준 (보통 width가 600px 이상이면 태블릿)
 const isTablet = width >= 600;
 
-const EthiopiaImage = () => {
-  return <StyledImage source={EthiopiaPng} />;
+const BrazilImage = () => {
+  return <StyledImage source={BrazilPng} />;
 };
 
 const StyledImage = styled.Image`
@@ -25,4 +25,4 @@ const StyledImage = styled.Image`
   height: ${isTablet ? responsiveHeight(351) : responsiveHeight(283)}px;
 `;
 
-export default EthiopiaImage;
+export default BrazilImage;
