+++ CaffeineDrop/components/eventPage/Card_CostaRica.js
@@ -1,12 +1,12 @@
 import React from "react";
 import { Image } from "react-native";
 import styled from "styled-components/native";
-import ColombiaPng from "../assets/EventPage/SouthAmerica2.png";
+import CostaRicaPng from "../../assets/EventPage/CentralAmerica2.png";
 import {
   responsiveFontSize,
   responsiveWidth,
   responsiveHeight,
-} from "../utils/responsive";
+} from "../../utils/responsive";
 
 import { Dimensions } from "react-native";
 
@@ -16,8 +16,8 @@ const { width, height } = Dimensions.get("window");
 // 태블릿 판별 기준 (보통 width가 600px 이상이면 태블릿)
 const isTablet = width >= 600;
 
-const ColombiaImage = () => {
-  return <StyledImage source={ColombiaPng} />;
+const CostaRicaImage = () => {
+  return <StyledImage source={CostaRicaPng} />;
 };
 
 const StyledImage = styled.Image`
@@ -25,4 +25,4 @@ const StyledImage = styled.Image`
   height: ${isTablet ? responsiveHeight(351) : responsiveHeight(283)}px;
 `;
 
-export default ColombiaImage;
+export default CostaRicaImage;
