+++ CaffeineDrop/components/eventPage/Drip.js
@@ -1,12 +1,12 @@
 import React from "react";
 import { Image } from "react-native";
 import styled from "styled-components/native";
-import VietnamPng from "../assets/EventPage/Asia1.png";
+import DripPng from "../../assets/EventPage/Drip.png";
 import {
   responsiveFontSize,
   responsiveWidth,
   responsiveHeight,
-} from "../utils/responsive";
+} from "../../utils/responsive";
 
 import { Dimensions } from "react-native";
 
@@ -16,13 +16,13 @@ const { width, height } = Dimensions.get("window");
 // 태블릿 판별 기준 (보통 width가 600px 이상이면 태블릿)
 const isTablet = width >= 600;
 
-const VietnamImage = () => {
-  return <StyledImage source={VietnamPng} />;
+const DripImage = () => {
+  return <StyledImage source={DripPng} />;
 };
 
 const StyledImage = styled.Image`
-  width: ${responsiveWidth(185)}px;
-  height: ${isTablet ? responsiveHeight(351) : responsiveHeight(283)}px;
+  width: ${responsiveWidth(342.938)}px;
+  height: ${isTablet ? responsiveHeight(444.438) : responsiveHeight(342.438)}px;
 `;
 
-export default VietnamImage;
+export default DripImage;
