+++ CaffeineDrop/components/eventPage/Loading2.js
@@ -4,7 +4,7 @@ import {
   responsiveFontSize,
   responsiveWidth,
   responsiveHeight,
-} from "../utils/responsive";
+} from "../../utils/responsive";
 
 const Loading2 = () => {
   return (
