+++ CaffeineDrop/components/eventPage/Loading3.js
@@ -4,7 +4,7 @@ import {
   responsiveFontSize,
   responsiveWidth,
   responsiveHeight,
-} from "../utils/responsive";
+} from "../../utils/responsive";
 
 const Loading3 = () => {
   return (
