+++ CaffeineDrop/components/eventPage/LoadingImg.js
@@ -1,12 +1,12 @@
 import React from "react";
 import { Image } from "react-native";
 import styled from "styled-components/native";
-import Loadingpng from "../assets/EventPage/Loading_Img.png";
+import Loadingpng from "../../assets/EventPage/Loading_Img.png";
 import {
   responsiveFontSize,
   responsiveWidth,
   responsiveHeight,
-} from "../utils/responsive";
+} from "../../utils/responsive";
 
 const LoadingImg = () => {
   return <StyledImage source={Loadingpng} />;
